+++ tools/sdk-cli/src/command_handlers.rs
@@ -372,12 +372,12 @@ pub(crate) async fn handle_command(
                 .await?;
             serde_json::to_string_pretty(&res).map_err(|e| e.into())
         }
-        Commands::NodeCredentials {} => match sdk()?.node_credentials()? {
+        Commands::NodeCredentials {} => match sdk()?.node_credentials().await? {
             Some(credentials) => serde_json::to_string_pretty(&credentials).map_err(|e| e.into()),
             None => Ok("No credentials".into()),
         },
         Commands::NodeInfo {} => {
-            serde_json::to_string_pretty(&sdk()?.node_info()?).map_err(|e| e.into())
+            serde_json::to_string_pretty(&sdk()?.node_info().await?).map_err(|e| e.into())
         }
         Commands::ConfigureNode { close_to_address } => {
             sdk()?
